magit-process: Add prompt regexps for GnuPG.
Fixes <https://github.com/magit/magit/issues/4076>.

When using Magit on a remote Git repository via TRAMP (using SSH), the
gpg-agent of the remote may prompt for a password.  Pinentry displays
the prompt through the terminal of the remote process, which until now
was not being handled by magit-process.

* lisp/magit-process.el (magit-process-password-prompt-regexps): Add
regexps to match the GnuPG prompts when requesting a password for a PGP
or SSH key through pinentry.